refactor(memory-sdk): isolate conformance scenarios
